refactor(runtime): extract context-budget shared helpers into a leaf module
Move the 15 cross-block shared pure helpers from context-budget.ts into
context-budget-helpers.ts so the upcoming synthesis-cache.ts and
history-compact.ts siblings can reuse them without a reverse import into
context-budget.ts.

Helpers moved: sha256, stableStringify, uniqueSorted, estimateTokens,
turnKey, groupEventsByTurn, finitePositive, nonEmpty, allNonEmpty,
increment, escapeAttribute, tokenizeSearchQuery, stableJsonLength,
estimateRuntimeEventChars, estimateRuntimeEventsTokens.

Behavior-neutral: context-budget.ts imports the helpers and re-exports
estimateTokens/estimateRuntimeEventsTokens to preserve public imports.

Part of #1084 (runtime/compaction lane, PR1).
